pyproject.toml

[build-system]
requires = ["setuptools >= 65.5.0"]
build-backend = "setuptools.build_meta"

[project]
name = "CC26"
dynamic = ["version"]
requires-python = ">= 3.12"
authors = [ # put your name, email here!
    {name = "Gavin Travis", email = "gavin.travis@wsu.edu"},
]
dependencies = [
    "requests",
    "flask",
    "google-auth",
    "google-auth-oauthlib",
    "google-api-python-client",
    "google-auth-httplib2",
]

[project.optional-dependencies]
dev = [
    "pre-commit",
    "ruff",
    "black",
    "isort",
    "pyright",
    "pytest",
]

[project.scripts]

[tool.setuptools.packages.find]
where = ["."]

[tool.black]
line-length = 80
[tool.ruff]
line-length = 80 # max line length
select = [
    "E", # code style errs
    "W", # code style warnings
    "F", # pyflakes issues -> fancy set of rules
    "D", # doc string rules
    "N", # pep8 naming conventions
    "B", # bugbear errs -> more fancy rules
]
ignore = [
    #"put err code here", # ignore specific err codes here, only do if its very unreasonable!
]
exclude = [".venv", "*.md", ".egg*", ".git", ".gitignore", "tests/**"] # stuff not to lint + format

[tool.pyright]
venvPath = "."
venv = ".venv/"
